Move FileIO props conversion into object_store_factory
